# cosmo: remove reflect calls in LambdaCDM benchmarks