This CL modifies the LambdaCDM benchmarks to use method-closures instead
of reflect to call specific LambdaCDM methods.

The gains are not super impressive for all cases but this makes the
benchmarks somewhat more correct.

```
$> benchstat ref-LambdaCDM.txt new.txt
name                                    old time/op  new time/op  delta
LambdaCDMEN-4                           60.5µs ± 2%  60.6µs ± 1%     ~     (p=0.351 n=20+20)
LambdaCDMENdistance-4                   8.09ms ± 7%  0.18ms ± 1%  -97.80%  (p=0.000 n=19+20)
LambdaCDME-4                            6.41ns ± 1%  6.41ns ± 1%     ~     (p=0.486 n=19+20)
LambdaCDMEinv-4                         11.7ns ± 1%  11.8ns ± 3%   +0.63%  (p=0.014 n=19+20)
LambdaCDMComovingDistance-4              210µs ± 2%   206µs ± 1%   -1.96%  (p=0.000 n=20+19)
LambdaCDMComovingTransverseDistance-4    211µs ± 1%   207µs ± 1%   -1.59%  (p=0.000 n=19+19)
LambdaCDMLuminosityDistance-4            213µs ± 2%   207µs ± 1%   -2.93%  (p=0.000 n=19+20)
LambdaCDMLookbackTime-4                  214µs ± 1%   213µs ± 1%   -0.66%  (p=0.000 n=16+19)
LambdaCDMNComovingDistance-4             2.09s ± 0%   2.07s ± 1%   -1.41%  (p=0.000 n=16+20)
LambdaCDMNLuminosityDistance-4           2.09s ± 1%   2.07s ± 1%   -0.99%  (p=0.000 n=20+20)
LambdaCDMNE-4                           8.22ms ± 7%  0.18ms ± 1%  -97.84%  (p=0.000 n=20+20)
LambdaCDMComovingDistanceOM-4            939ns ±12%   161ns ± 1%  -82.85%  (p=0.000 n=20+20)
LambdaCDMLookbackTimeOM-4               1.32µs ± 9%  0.49µs ± 1%  -62.86%  (p=0.000 n=20+20)
LambdaCDMComovingDistanceFlat-4         1.42µs ± 7%  0.62µs ± 1%  -56.55%  (p=0.000 n=20+20)
LambdaCDMComovingDistancePositiveOk0-4   210µs ± 1%   207µs ± 2%   -1.39%  (p=0.000 n=20+20)
```